feat(download.py): add download_yt_video function to download YouTube video using subprocess
The download.py file was added with a new function called download_yt_video. This function uses the subprocess module to run a command that downloads a YouTube video using the Youtube-Upload/youtube_auth.js script. If an error occurs during the download process, it will be caught and printed.