docs(README.md): update instructions for OAuth consent screen and uploading a file
The instructions for the OAuth consent screen have been updated to clarify that test users should be added. Additionally, a note has been added to specify that only test users with a Gmail account can sign in to the YouTube v3 API.

The instructions for uploading a file have been updated to remove an unnecessary line break.

A note has been added to suggest trying the script in a Linux environment and installing the dependencies manually if there are any errors when installing requirements.txt.

Author: Dikshit Singh